+++ test/RedemptionIdle.t.sol
@@ -391,7 +391,7 @@ contract RedemptionIdleTest is Test {
     }
 
     function testSetRedemptionFeeSameFail() public {
-        uint96 oldFee = redemption.redemptionFee();
+        uint256 oldFee = redemption.redemptionFee();
 
         hoax(owner);
         vm.expectRevert(IRedemption.BadArgs.selector);
@@ -399,7 +399,7 @@ contract RedemptionIdleTest is Test {
     }
 
     function testCalculateFee() public {
-        uint96 fee = 5; // 0.05%
+        uint256 fee = 5; // 0.05%
         hoax(owner);
         redemption.setRedemptionFee(fee);
 
@@ -410,7 +410,7 @@ contract RedemptionIdleTest is Test {
     }
 
     function testCalculateUstbInWithFee() public {
-        uint96 fee = 5; // 0.05%
+        uint256 fee = 5; // 0.05%
         hoax(owner);
         redemption.setRedemptionFee(fee);
 
@@ -422,7 +422,7 @@ contract RedemptionIdleTest is Test {
     }
 
     function testRedeemWithFee() public {
-        uint96 fee = 5; // 0.05%
+        uint256 fee = 5; // 0.05%
         hoax(owner);
         redemption.setRedemptionFee(fee);
